LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/tables/form-in-tables-expected.txt


FAIL Computed display of form inside TABLE should be 'none' assert_equals: expected "none" but got "block"
FAIL Computed display of form inside THEAD should be 'none' assert_equals: expected "none" but got "block"
FAIL Computed display of form inside TBODY should be 'none' assert_equals: expected "none" but got "block"
FAIL Computed display of form inside TFOOT should be 'none' assert_equals: expected "none" but got "block"
FAIL Computed display of form inside TR should be 'none' assert_equals: expected "none" but got "block"
FAIL Computed display of form inside TABLE should be 'none' (!important UA style)) assert_equals: expected "none" but got "block"
FAIL Computed display of form inside THEAD should be 'none' (!important UA style)) assert_equals: expected "none" but got "block"
FAIL Computed display of form inside TBODY should be 'none' (!important UA style)) assert_equals: expected "none" but got "block"
FAIL Computed display of form inside TFOOT should be 'none' (!important UA style)) assert_equals: expected "none" but got "block"
FAIL Computed display of form inside TR should be 'none' (!important UA style)) assert_equals: expected "none" but got "block"
